requirements.txt
anyio==4.6.0
certifi==2024.8.30
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
psycopg2==2.9.9
python-telegram-bot==21.6
sniffio==1.3.1
